fix: components now show updated amount following a parrtial payment
