+++ src/cartridges/int_adyen_SFRA/cartridge/client/default/js/adyen_checkout/helpers.js
@@ -49,32 +49,6 @@ function paymentFromComponent(data, component = {}) {
   });
 }
 
-function makePartialPayment(requestData) {
-  let error;
-  $.ajax({
-    url: 'Adyen-partialPayment',
-    type: 'POST',
-    data: JSON.stringify(requestData),
-    contentType: 'application/json; charset=utf-8',
-    async: false,
-    success(response) {
-      if (response.error) {
-        error = {
-          error: true,
-        };
-      } else {
-        const { giftCards, ...rest } = response;
-        store.checkout.options.amount = rest.remainingAmount;
-        store.partialPaymentsOrderObj = rest;
-        sessionStorage.setItem('partialPaymentsObj', JSON.stringify(rest));
-        store.addedGiftCards = giftCards;
-        setOrderFormData(response);
-      }
-    },
-  }).fail(() => {});
-  return error;
-}
-
 function resetPaymentMethod() {
   $('#requiredBrandCode').hide();
   $('#selectedIssuer').val('');
@@ -162,5 +136,4 @@ module.exports = {
   showValidation,
   createShowConfirmationForm,
   getInstallmentValues,
-  makePartialPayment,
 };
