+++ src/cartridges/int_adyen_SFRA/cartridge/client/default/js/adyen_checkout/makePartialPayment.js
@@ -1,0 +1,34 @@
+const store = require('../../../../store');
+const { renderGenericComponent } = require('./renderGenericComponent');
+const helpers = require('./helpers');
+
+function makePartialPayment(requestData) {
+  let error;
+  $.ajax({
+    url: 'Adyen-partialPayment',
+    type: 'POST',
+    data: JSON.stringify(requestData),
+    contentType: 'application/json; charset=utf-8',
+    async: false,
+    success(response) {
+      if (response.error) {
+        error = {
+          error: true,
+        };
+      } else {
+        const { giftCards, ...rest } = response;
+        store.checkout.options.amount = rest.remainingAmount;
+        store.partialPaymentsOrderObj = rest;
+        sessionStorage.setItem('partialPaymentsObj', JSON.stringify(rest));
+        store.addedGiftCards = giftCards;
+        helpers.setOrderFormData(response);
+        renderGenericComponent(true);
+      }
+    },
+  }).fail(() => {});
+  return error;
+}
+
+module.exports = {
+  makePartialPayment,
+};
