+++ src/cartridges/int_adyen_SFRA/cartridge/client/default/js/adyen_checkout/renderGenericComponent.js
@@ -62,8 +62,8 @@ function renderGiftCardLogo(imagePath) {
     headingImg.src = `${imagePath}genericgiftcard.png`;
   }
 }
-
-function applyGiftCards() {
+// eslint-disable-next-line complexity
+function applyGiftCards(refresh) {
   const now = new Date().toISOString();
   const { amount } = store.checkoutConfiguration;
   const { orderAmount } = store.partialPaymentsOrderObj;
@@ -72,8 +72,9 @@ function applyGiftCards() {
     (cart) => now > cart.expiresAt,
   );
   const isCartModified =
-    amount.currency !== orderAmount.currency ||
-    amount.value !== orderAmount.value;
+    refresh === false &&
+    (amount.currency !== orderAmount.currency ||
+      amount.value !== orderAmount.value);
 
   if (isPartialPaymentExpired) {
     removeGiftCards();
@@ -186,64 +187,66 @@ function setGiftCardContainerVisibility() {
 /**
  * Calls createSession and then renders the retrieved payment methods (including card component)
  */
-module.exports.renderGenericComponent =
-  async function renderGenericComponent() {
-    if (Object.keys(store.componentsObj).length !== 0) {
-      await unmountComponents();
-    }
+module.exports.renderGenericComponent = async function renderGenericComponent(
+  refresh = false,
+) {
+  if (Object.keys(store.componentsObj).length !== 0) {
+    await unmountComponents();
+  }
 
-    const session = await createSession();
-    const giftCardsData = await fetchGiftCards();
+  const session = await createSession();
+  const giftCardsData = await fetchGiftCards();
 
-    store.checkoutConfiguration.session = {
-      id: session.id,
-      sessionData: session.sessionData,
-      imagePath: session.imagePath,
-      adyenDescriptions: session.adyenDescriptions,
-    };
-    store.checkout = await AdyenCheckout(store.checkoutConfiguration);
-
-    setGiftCardContainerVisibility();
-    const { totalDiscountedAmount, giftCards } = giftCardsData;
-    store.addedGiftCards = giftCards;
-    if (giftCards?.length) {
-      const lastGiftCard = giftCards[store.addedGiftCards.length - 1];
-      store.partialPaymentsOrderObj = giftCardsData.giftCards?.length
-        ? { ...lastGiftCard, totalDiscountedAmount }
-        : null;
-    }
-
-    setCheckoutConfiguration(store.checkout.options);
-    setInstallments(store.checkout.options.amount);
-    setAmazonPayConfig(store.checkout.paymentMethodsResponse);
-    document.querySelector('#paymentMethodsList').innerHTML = '';
-
-    renderStoredPaymentMethods(
-      store.checkout.paymentMethodsResponse,
-      session.imagePath,
-    );
-    renderPaymentMethods(
-      store.checkout.paymentMethodsResponse,
-      session.imagePath,
-      session.adyenDescriptions,
-    );
-    renderPosTerminals(session.adyenConnectedTerminals);
-
-    renderGiftCardLogo(session.imagePath);
-
-    if (store.addedGiftCards?.length) {
-      applyGiftCards();
-    }
-
-    attachGiftCardAddButtonListener();
-
-    const firstPaymentMethod = document.querySelector(
-      'input[type=radio][name=brandCode]',
-    );
-    firstPaymentMethod.checked = true;
-    helpers.displaySelectedMethod(firstPaymentMethod.value);
-
-    helpers.createShowConfirmationForm(
-      window.ShowConfirmationPaymentFromComponent,
-    );
+  store.checkoutConfiguration.session = {
+    id: session.id,
+    sessionData: session.sessionData,
+    imagePath: session.imagePath,
+    adyenDescriptions: session.adyenDescriptions,
   };
+  store.checkout = await AdyenCheckout(store.checkoutConfiguration);
+
+  setGiftCardContainerVisibility();
+  const { totalDiscountedAmount, giftCards } = giftCardsData;
+  store.addedGiftCards = giftCards;
+  if (giftCards?.length) {
+    const lastGiftCard = giftCards[store.addedGiftCards.length - 1];
+    store.partialPaymentsOrderObj = giftCardsData.giftCards?.length
+      ? { ...lastGiftCard, totalDiscountedAmount }
+      : null;
+  }
+
+  setCheckoutConfiguration(store.checkout.options);
+  setInstallments(store.checkout.options.amount);
+  setAmazonPayConfig(store.checkout.paymentMethodsResponse);
+  document.querySelector('#paymentMethodsList').innerHTML = '';
+
+  renderStoredPaymentMethods(
+    store.checkout.paymentMethodsResponse,
+    session.imagePath,
+  );
+
+  renderPaymentMethods(
+    store.checkout.paymentMethodsResponse,
+    session.imagePath,
+    session.adyenDescriptions,
+  );
+  renderPosTerminals(session.adyenConnectedTerminals);
+
+  renderGiftCardLogo(session.imagePath);
+
+  if (store.addedGiftCards?.length) {
+    applyGiftCards(refresh);
+  }
+
+  attachGiftCardAddButtonListener();
+
+  const firstPaymentMethod = document.querySelector(
+    'input[type=radio][name=brandCode]',
+  );
+  firstPaymentMethod.checked = true;
+  helpers.displaySelectedMethod(firstPaymentMethod.value);
+
+  helpers.createShowConfirmationForm(
+    window.ShowConfirmationPaymentFromComponent,
+  );
+};
