+++ src/cartridges/int_adyen_overlay/cartridge/scripts/adyenSessions.js
@@ -37,13 +37,23 @@ function createSession(basket, customer, countryCode) {
 
     // There is no basket for myAccount session requests
     if(basket) {
+        AdyenLogs.error_log(`session.privacy.giftCardResponse ${session.privacy.giftCardResponse}`);
+
+        let amount;
+        if(session.privacy.giftCardResponse && JSON.parse(session.privacy.giftCardResponse).remainingAmount) {
+            AdyenLogs.error_log(`inside if statement`);
+            amount = JSON.parse(session.privacy.giftCardResponse).remainingAmount
+        } else {
+            AdyenLogs.error_log(`inside else statement`);
+            amount = {
+                currency: basket.currencyCode,
+                value: AdyenHelper.getCurrencyValueForApi(basket.getTotalGrossPrice()).value
+            };
+        }
       //TODO: Change AdyenHelper so that this object can have a different name. Not creating a payment request here
       let paymentRequest = {
         merchantAccount: AdyenConfigs.getAdyenMerchantAccount(),
-        amount: {
-          currency: basket.currencyCode,
-          value: AdyenHelper.getCurrencyValueForApi(basket.getTotalGrossPrice()).value,
-        },
+        amount,
       };
 
       // Add Risk data
